+++ scripts/build.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
+# usage: build.sh <src_dir> <build_dir>
 for f in $(find $1 -name '*.tex'); do
-    echo $f
     OUT_DIR=$2/$(dirname "${f#./}")
-    echo "mkdir -p $OUT_DIR"
-    echo "latexmk -output-directory=$OUT_DIR -pdf $f"
-    echo "latexmk -c -output-directory=$OUT_DIR $f"
+    mkdir -p $OUT_DIR
+    latexmk -output-directory=$OUT_DIR -pdf $f
+    latexmk -c -output-directory=$OUT_DIR $f
 done
+./scripts/gen_filetree.py $2
